[ci] Remove usage of build.py in favor of setup.py (#3537)
* Remove usage of build.py in favor of setup.py

* fix make_changelog

* fetch git history

* add a script upload_reelase.py
