+++ .github/workflows/presubmit.yml
@@ -89,46 +89,46 @@ jobs:
           git diff _fake_squash _enforced_format --exit-code
 
   check_static_analyzer:
-      name: Check Static Analyzer
-      runs-on: ubuntu-latest
-      needs: check_files
-      if: needs.check_files.outputs.run_job == 'true'
-      steps:
-        - uses: actions/checkout@v2
-          with:
-            submodules: "recursive"
+    name: Check Static Analyzer
+    runs-on: ubuntu-latest
+    needs: check_files
+    if: needs.check_files.outputs.run_job == 'true'
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: "recursive"
 
-        - name: Set up Python ${{ matrix.python-version }}
-          uses: actions/setup-python@v2
-          with:
-            python-version: 3.8
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
 
-        - name: Pylint
-          run: |
-            python3 -m pip install --user pylint
-            # Make sure pylint doesn't regress
-            pylint python/taichi/ --disable=all --enable=$(python scripts/generate_pylint_tags.py)
-            if [ $? -eq 0 ]
-            then
-              echo "PASSED: pylint is happy"
-              exit 0
-            else
-              echo "FAILED: please run the pylint command above and make sure it passes"
-              exit 1
-            fi
+      - name: Pylint
+        run: |
+          python3 -m pip install --user pylint
+          # Make sure pylint doesn't regress
+          pylint python/taichi/ --disable=all --enable=$(python scripts/generate_pylint_tags.py)
+          if [ $? -eq 0 ]
+          then
+            echo "PASSED: pylint is happy"
+            exit 0
+          else
+            echo "FAILED: please run the pylint command above and make sure it passes"
+            exit 1
+          fi
 
-        - name: clang-tidy
-          run: |
-            # https://docs.github.com/en/packages/managing-github-packages-using-github-actions-workflows/publishing-and-installing-a-package-with-github-actions#upgrading-a-workflow-that-accesses-ghcrio
-            echo $CR_PAT | docker login ghcr.io -u ${{ github.actor }} --password-stdin
-            docker pull ghcr.io/taichi-dev/taichidev-cpu-ubuntu18.04:v0.1.0
-            docker run -id --user dev --name check_clang_tidy ghcr.io/taichi-dev/taichidev-cpu-ubuntu18.04:v0.1.0 /bin/bash
-            tar -cf - ../${{ github.event.repository.name }} --mode u=+rwx,g=+rwx,o=+rwx --owner dev --group dev | docker cp - check_clang_tidy:/home/dev/
-            docker exec --user root check_clang_tidy apt install -y clang-tidy-10
-            docker exec --user dev check_clang_tidy /home/dev/taichi/.github/workflows/scripts/check_clang_tidy.sh "$CI_SETUP_CMAKE_ARGS"
-          env:
-            CR_PAT: ${{ secrets.GITHUB_TOKEN }}
-            CI_SETUP_CMAKE_ARGS: -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CC:BOOL=ON -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=OFF
+      - name: clang-tidy
+        run: |
+          # https://docs.github.com/en/packages/managing-github-packages-using-github-actions-workflows/publishing-and-installing-a-package-with-github-actions#upgrading-a-workflow-that-accesses-ghcrio
+          echo $CR_PAT | docker login ghcr.io -u ${{ github.actor }} --password-stdin
+          docker pull ghcr.io/taichi-dev/taichidev-cpu-ubuntu18.04:v0.1.0
+          docker run -id --user dev --name check_clang_tidy ghcr.io/taichi-dev/taichidev-cpu-ubuntu18.04:v0.1.0 /bin/bash
+          tar -cf - ../${{ github.event.repository.name }} --mode u=+rwx,g=+rwx,o=+rwx --owner dev --group dev | docker cp - check_clang_tidy:/home/dev/
+          docker exec --user root check_clang_tidy apt install -y clang-tidy-10
+          docker exec --user dev check_clang_tidy /home/dev/taichi/.github/workflows/scripts/check_clang_tidy.sh "$CI_SETUP_CMAKE_ARGS"
+        env:
+          CR_PAT: ${{ secrets.GITHUB_TOKEN }}
+          CI_SETUP_CMAKE_ARGS: -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CC:BOOL=ON -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=OFF
 
   build_and_test_cpu_linux:
     name: Build and Test linux (CPU)
@@ -218,9 +218,13 @@ jobs:
 
       - name: Build & Install
         run: |
+          if [[ ${{needs.check_files.outputs.run_job}} == false ]]; then
+            exit 0
+          fi
+          export PATH=`pwd`/taichi-llvm/bin/:$PATH
           .github/workflows/scripts/unix_build.sh
         env:
-          CI_SETUP_CMAKE_ARGS: -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CC:BOOL=${{ matrix.with_cc }} -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=${{ matrix.with_cpp_tests }}
+          TAICHI_CMAKE_ARGS: -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CC:BOOL=${{ matrix.with_cc }} -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=${{ matrix.with_cpp_tests }}
           CXX: clang++
       # [DEBUG] Copy this step around to enable debugging inside Github Action instances.
       #- name: Setup tmate session
@@ -339,7 +343,7 @@ jobs:
           rm -rf $HOME/Library/Python/3.8/lib/python/site-packages/taichi
           .github/workflows/scripts/unix_build.sh
         env:
-          CI_SETUP_CMAKE_ARGS: -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CUDA:BOOL=OFF -DTI_WITH_CC:BOOL=OFF -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=ON
+          TAICHI_CMAKE_ARGS: -DTI_WITH_OPENGL:BOOL=OFF -DTI_WITH_CUDA:BOOL=OFF -DTI_WITH_CC:BOOL=OFF -DTI_WITH_VULKAN:BOOL=OFF -DTI_BUILD_TESTS:BOOL=ON
           CXX: clang++
 
       - name: Test
