+++ requirements_dev.txt
@@ -11,3 +11,5 @@ sourceinspect
 isort
 pylint
 requests==2.26
+twine
+wheel
